update has or not obj for all grids

+++ yolo/yolo_v1.py
@@ -211,9 +211,9 @@ class YOLO_V1_Loss(nn.Module):
         # mse loss function
         mse=nn.MSELoss(reduction='none')
 
-        # if has obj or not
-        lambda_obj=target_conf>0.9
-        lambda_noobj=target_conf<0.1
+        # if has obj or not for all grids
+        lambda_obj=(target_conf>0.9).float()
+        lambda_noobj=(target_conf<0.1).float()
 
         # coordinate loss
         # print(f'pred_bbox.shape:{pred_bbox.shape}, target_bbox.shape:{target_bbox.shape}')
